back-end may finish

+++ code/back-end/ffms/src/main/java/org/csu/ffms/domain/Income.java
@@ -7,7 +7,7 @@ import java.util.Date;
 @Component("Income")
 public class Income {
     private int incomeId;
-    private int income;
+    private float income;
     private String userId;
     private Date time;
     private String description;
@@ -21,11 +21,11 @@ public class Income {
         this.incomeId = incomeId;
     }
 
-    public int getIncome() {
+    public float getIncome() {
         return income;
     }
 
-    public void setIncome(int income) {
+    public void setIncome(float income) {
         this.income = income;
     }
